requirements.txt
beautifulsoup4
requests
Pillow
ImageHash